+++ app/scripts/configs/options-info.js
@@ -127,6 +127,14 @@ export const OPTIONS_INFO = {
     name: 'Sort Largest On Top',
     inlineOptions: YES_NO
   },
+  showTexts: {
+    name: 'Show texts',
+    inlineOptions: YES_NO
+  },
+  staggered: {
+    name: 'Staggered',
+    inlineOptions: YES_NO
+  },
   minSquareSize: {
     name: 'Minimum size',
     inlineOptions: AVAILABLE_WIDTHS_AND_NONE,
@@ -493,6 +501,18 @@ export const OPTIONS_INFO = {
       10: { name: '10px', value: 10 },
       12: { name: '12px', value: 12 },
       16: { name: '16px', value: 16 },
+    }
+  },
+
+
+  annotationHeight: {
+    name: 'Annotation Height',
+    inlineOptions: {
+      8: { name: '8px', value: 8 },
+      10: { name: '10px', value: 10 },
+      12: { name: '12px', value: 12 },
+      16: { name: '16px', value: 16 },
+      20: { name: '20px', value: 20 }
     }
   },